fixing format to pass travis_ci test

+++ python/PDFs/physics/Amp3Body.cpp
@@ -47,8 +47,8 @@ void init_Amp3Body(py::module &m) {
 
             py::array_t<fptype> pyparticles{{(size_t)3, 3 * numEvents}};
             py::array_t<fptype> pyvariables{{(size_t)3, numEvents}};
-            py::array_t<fptype> pyweights{numEvents};
-            py::array_t<fptype> pyflags{numEvents};
+            py::array_t<fptype> pyweights{static_cast<ssize_t>(numEvents)};
+            py::array_t<fptype> pyflags{static_cast<ssize_t>(numEvents)};
 
             for(int i = 0; i < 3; i++) {
                 for(int j = 0, k = 0; j < numEvents; j++, k = k + 3) {
